Refactored Cycle to eliminate volumeId, since it's not needed.

+++ app/src/main/java/com/ssimon/cyclesactivity/Const.java
@@ -13,6 +13,6 @@ package com.ssimon.cyclesactivity;
 
 public interface Const {
     int NULL_RESOURCE_ID = 0;
-    int MIN_DATABASE_ID = 1;
-    int UNSET_DATABASE_ID = MIN_DATABASE_ID - 1;
+    long MIN_DATABASE_ID = 1;
+    long UNSET_DATABASE_ID = MIN_DATABASE_ID - 1;
 }
